requirements.txt
torch
pandas
numpy
nest-asyncio
pandas
ipywidgets
stadle_client
streamlit
